Move the migration event adapter to a better place and mention in docs.
