+++ akka-cluster-sharding/src/main/scala/akka/cluster/sharding/internal/EventSourcedRememberEntitiesCoordinatorStore.scala
@@ -4,14 +4,14 @@
 
 package akka.cluster.sharding.internal
 
-import akka.actor.{ ActorLogging, Props }
+import akka.actor.ActorLogging
+import akka.actor.Props
 import akka.annotation.InternalApi
-import akka.cluster.sharding.{ ClusterShardingSerializable, ClusterShardingSettings }
 import akka.cluster.sharding.ShardCoordinator.Internal
-import akka.cluster.sharding.ShardCoordinator.Internal.ShardHomeAllocated
 import akka.cluster.sharding.ShardRegion.ShardId
+import akka.cluster.sharding.ClusterShardingSerializable
+import akka.cluster.sharding.ClusterShardingSettings
 import akka.persistence._
-import akka.persistence.journal.{ EventAdapter, EventSeq }
 
 import scala.collection.mutable
 
@@ -22,23 +22,6 @@ import scala.collection.mutable
 private[akka] object EventSourcedRememberEntitiesCoordinatorStore {
   def props(typeName: String, settings: ClusterShardingSettings): Props =
     Props(new EventSourcedRememberEntitiesCoordinatorStore(typeName, settings))
-
-  class FromOldCoordinatorState() extends EventAdapter {
-    override def manifest(event: Any): String =
-      ""
-
-    override def toJournal(event: Any): Any =
-      event
-
-    override def fromJournal(event: Any, manifest: String): EventSeq = {
-      event match {
-        case ShardHomeAllocated(shardId, _) =>
-          EventSeq.single(shardId)
-        case _ => EventSeq.empty
-      }
-
-    }
-  }
 
   case class State(shards: Set[ShardId], writtenMigrationMarker: Boolean = false) extends ClusterShardingSerializable
